requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0

sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
psycopg2-binary==2.9.9

yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4

pandas-ta @ git+https://github.com/aarigs/pandas-ta
vectorbt==0.26.0
numba==0.59.0

groq==0.5.0
instructor==1.0.0
pydantic==2.6.3

plotly==5.19.0
mplfinance==0.12.10b0
matplotlib

line-bot-sdk==3.7.0

streamlit>=1.30.0
beautifulsoup4
quantstats